[merge] Move most tables into merge.base module
